+++ apps/cloud/CHANGELOG.md
@@ -1,3 +1,9 @@
 # @xpert-ai/xpert-ui
 
+## 3.9.3
+
+### Patch Changes
+
+- 07057a6: pet of chatkit
+
 ## 3.9.0